refactor: front: Added redux, hook alias. Change folder and components file name.
