+++ frontend/src/services/redux/auth/index.ts
@@ -1,0 +1,2 @@
+export * from './auth.actions'
+export * from './auth.slice'
